# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    GenericError:
        description: GenericError generic error
        properties:
            message:
                description: message
                type: string
                x-go-name: Message
        type: object
        x-go-package: product-api/sdk/models
    Product:
        description: Product Product defines the structure for an API product
        properties:
            description:
                description: the description for this product
                type: string
                x-go-name: Description
            id:
                description: the id for this product
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: the name of this product
                type: string
                x-go-name: Name
            price:
                description: the price for this product
                format: float
                minimum: 0.01
                type: number
                x-go-name: Price
            sku:
                description: the SKU for this product
                pattern: '[a-z]+-[a-z]+-[a-z]+'
                type: string
                x-go-name: SKU
        required:
            - id
            - name
            - description
            - price
            - sku
        type: object
        x-go-package: product-api/sdk/models
    ValidationError:
        description: ValidationError validation error
        properties:
            messages:
                description: messages
                items:
                    type: string
                type: array
                x-go-name: Messages
        type: object
        x-go-package: product-api/sdk/models
info:
    description: Documentation for Product API
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            description: Adds a product to the database
            operationId: createProduct
            parameters:
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/Product'
            responses:
                "204":
                    $ref: '#/responses/created'
                "422":
                    $ref: '#/responses/errorValidation'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
        put:
            description: Updates a product
            operationId: updateProduct
            parameters:
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/Product'
            responses:
                "201":
                    $ref: '#/responses/noContent'
    /products/{id}:
        delete:
            operationId: deleteProduct
            parameters:
                - description: The id of the product for which the operation relates
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
                "404":
                    $ref: '#/responses/errorResponse'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
        get:
            description: Returns a single product
            operationId: listSingleProduct
            parameters:
                - description: The id of the product for which the operation relates
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/productResponse'
                "404":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
produces:
    - application/json
responses:
    created:
        description: ""
    errorResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    errorValidation:
        description: ""
        schema:
            $ref: '#/definitions/ValidationError'
    noContent:
        description: ""
    productResponse:
        description: ""
        schema:
            $ref: '#/definitions/Product'
    productsResponse:
        description: ""
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
swagger: "2.0"